requirements.txt
google-genai>=1.0.0
yt-dlp>=2024.10.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
# Local ASR — free, high-quality Arabic word-level timestamps. No API cost.
faster-whisper>=1.0.0
